tags instances after their certificate if signed
this prevents a replay attack by using an already signed / running aws
instance's ID to get a new certificate for another service.

to bypass, this would require an attacker to be able to:
* run a puppet agent in a vpc that can talk to your master
* reference a running instance id that is running your environment
 - that does not have a puppet_cert_signed tag

Obviously the first approach is to provision an instance in your AWS
environment / with an iam profile / root account. However if those
credentials are compromised already, that is a bigger issue. If this is
a risk for you organization, you may want to use a different autosign
procedure.